walk does not return
